Introduce `torch.distributed.nn.utils.describe()`
This PR introduces the `describe()` function that enables constructing `torch.nn.Module` instances on the meta device for inspection purposes. It is meant to be used if the size of a module is too big to fit on a single machine and you want to inspect it without instantiating.

Internally ``describe()`` forces all parameters, buffers, and other tensors within the scope of the module and its sub-modules to use the meta device regardless of the actual device type. The returned module can be used for inspection purposes and afterwards its ``materialize()`` method can be called to construct a fully-instantiated module.

```
>>> import torch.distributed.nn
>>> m = torch.distributed.nn.utils.describe(torch.nn.Linear, 5, 1)
>>> m.weight
Parameter containing:
tensor(..., device='meta', requires_grad=True)
>>> m = m.materialize()
>>> m.weight
Parameter containing:
tensor([[-1.4677e+24,  4.5915e-41,  1.4013e-45,  0.0000e+00,
             -1.4677e+24, 4.5915e-41]], requires_grad=True)
```

However note that ``describe()`` uses a "best effort" algorithm and is not guaranteed to succeed if the underlying module's implementation cannot be mapped to the meta device. If you are the module author, you can use the `is_described()` function to find out whether your module is being instantiated in the scope of a ``describe()`` call and use an alternate logic if required.

```
class MyModule(Module):
    def __init__(self):
        if torch.distributed.nn.utils.is_described():
            self.myparam = torch.empty([10,10], device="meta")
        else:
            self.myparam = load_myparam()
```



**describe() vs. skip_init()**

Although they sound similar `describe()` and `torch.nn.utils.skip_init()` serve different purposes. `describe()` is meant for lazy model materialization and targets model users, while ``skip_init()`` is meant for optimized module initialization and targets model authors.

Technical differences between the two functions are:

- `describe()` returns a module where all its parameters and buffers reside on the meta device, while `skip_init()` returns a module where some or all its parameters or buffers are empty tensors allocated on a real device.
- `describe()` does not require any code changes in the module, while `skip_init()` expects the module and its submodules to take a `device` argument. This means ``skip_init()`` is intrusive, but also guaranteed to work; on the other hand ``describe()`` can potentially fail to construct a module if the module initializes its state in an unconventional way that conflicts with the meta device.

Differential Revision: [D31504090](https://our.internmc.facebook.com/intern/diff/D31504090/)

ghstack-source-id: 140096191
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66317